refactor: 교사 채팅 및 리포트 표시 데이터 기준 정리
- 채팅방 상세 의도 태그를 intentType 기준으로 매핑
- 오타 필드 intentLavel 및 intentLabel 의존 로직 제거
- 리포트 날짜 시간 포맷 유틸을 공통 utils로 분리
- formatDateTime 표시 형식을 점 구분 날짜 형식으로 변경
