+++ src/features/teacher/reports/lib/reportFormatters.ts
@@ -14,27 +14,6 @@ export const formatDateOnly = (value: string) => {
   return `${year}-${month}-${day}`;
 };
 
-export const formatDateTime = (value?: string) => {
-  if (!value) {
-    return '-';
-  }
-
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) {
-    return value;
-  }
-
-  const year = parsed.getFullYear();
-  const month = `${parsed.getMonth() + 1}`.padStart(2, '0');
-  const day = `${parsed.getDate()}`.padStart(2, '0');
-  const hour24 = parsed.getHours();
-  const minute = `${parsed.getMinutes()}`.padStart(2, '0');
-  const period = hour24 >= 12 ? '오후' : '오전';
-  const hour12 = hour24 % 12 || 12;
-
-  return `${year}-${month}-${day} ${period} ${hour12}:${minute}`;
-};
-
 export const formatPreviewName = (name?: string) => {
   if (!name || !name.trim()) {
     return '-';
